+++ test/writer.test.js
@@ -8,6 +8,35 @@ var VERSION = '1.2.3';
 
 describe('writer', function () {
 
+  describe('getCommitUrl', function () {
+
+    it('makes a valid URL for a BitBucket repository', function () {
+      var url = 'https://bitbucket.org/lob/generate-changelog';
+      var commitHash = '1234567890';
+
+      var linkUrl = Writer.getCommitUrl(url, commitHash);
+      Expect(linkUrl).to.equal(url + '/commits/' + commitHash);
+    });
+
+    it('makes a valid URL for a GitHub repository', function () {
+      var url = 'https://github.com/lob/generate-changelog';
+      var commitHash = '1234567890';
+
+      var linkUrl = Writer.getCommitUrl(url, commitHash);
+      Expect(linkUrl).to.equal(url + '/commit/' + commitHash);
+    });
+
+    it('makes a valid URL for a Gitlab repository', function () {
+      var url = 'https://gitlab.com/lob/generate-changelog.git';
+      var expectedUrl = 'https://gitlab.com/lob/generate-changelog';
+      var commitHash = '1234567890';
+
+      var linkUrl = Writer.getCommitUrl(url, commitHash);
+      Expect(linkUrl).to.equal(expectedUrl + '/commit/' + commitHash);
+    });
+
+  });
+
   describe('markdown', function () {
 
     it('makes heading h2 if major version', function () {